+++ composeApp/src/androidMain/kotlin/com/arn/scrobble/utils/AndroidStuff.kt
@@ -8,7 +8,6 @@ import android.content.Context
 import android.media.MediaMetadata
 import android.os.Build
 import android.os.Bundle
-import android.provider.Settings
 import android.view.View
 import android.view.ViewTreeObserver
 import android.widget.Toast
@@ -18,7 +17,6 @@ import androidx.annotation.StringRes
 import androidx.datastore.core.DataStoreFactory
 import co.touchlab.kermit.Logger
 import com.arn.scrobble.BuildKonfig
-import com.arn.scrobble.media.NLService
 import com.arn.scrobble.pref.WidgetPrefs
 import com.arn.scrobble.pref.WidgetPrefsSerializer
 import com.arn.scrobble.utils.Stuff.SCROBBLER_PROCESS_NAME
@@ -209,24 +207,6 @@ object AndroidStuff {
                 }
             }
         )
-    }
-
-
-    fun isNotificationListenerEnabled(): Boolean {
-        // adapted from NotificationManagerCompat.java
-
-        val enabledNotificationListeners = try {
-            Settings.Secure.getString(
-                applicationContext.contentResolver,
-                "enabled_notification_listeners"
-            )
-        } catch (e: SecurityException) {
-            return false
-        }
-
-        val nlsComponentStr = "${applicationContext.packageName}/${NLService::class.qualifiedName}"
-        // check for the exact component name instead of just package name
-        return enabledNotificationListeners?.split(":")?.any { it == nlsComponentStr } == true
     }
 
 }