+++ composeApp/src/commonMain/kotlin/com/arn/scrobble/main/ScrobblerStatus.kt
@@ -27,6 +27,7 @@ sealed interface ScrobblerState {
         val reason: String,
         val subReason: String,
         val desc: String,
+        val rssMb: Int,
         val isProbablySystemKill: Boolean
     ) {
         fun formatted() = "$desc ($reason) $subReason".trimEnd()