+++ composeApp/src/commonMain/kotlin/com/arn/scrobble/pref/ProxyPrefDialog.kt
@@ -1,5 +1,6 @@
 package com.arn.scrobble.pref
 
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -17,6 +18,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
 import com.arn.scrobble.ui.OutlinedToggleButtons
 import com.arn.scrobble.ui.PanoOutlinedTextField
@@ -29,31 +31,36 @@ import okhttp3.HttpUrl
 import org.jetbrains.compose.resources.stringResource
 import pano_scrobbler.composeapp.generated.resources.Res
 import pano_scrobbler.composeapp.generated.resources.host
+import pano_scrobbler.composeapp.generated.resources.password
 import pano_scrobbler.composeapp.generated.resources.port
 import pano_scrobbler.composeapp.generated.resources.proxy_socks5
 import pano_scrobbler.composeapp.generated.resources.system
+import pano_scrobbler.composeapp.generated.resources.username
 
 @Composable
 fun ProxyPrefDialog(modifier: Modifier = Modifier) {
     Column(
+        verticalArrangement = Arrangement.spacedBy(8.dp),
         modifier = modifier
     ) {
-        val enabled by PlatformStuff.mainPrefs.data.collectAsStateWithInitialValue { it.customProxyEnabled }
-        val host by PlatformStuff.mainPrefs.data.collectAsStateWithInitialValue { it.proxyHost }
-        val port by PlatformStuff.mainPrefs.data.collectAsStateWithInitialValue { it.proxyPort }
+        val proxy by PlatformStuff.mainPrefs.data.collectAsStateWithInitialValue { it.proxy }
 
-        var enabledEditable by remember { mutableStateOf(enabled) }
-        var hostEditable by remember { mutableStateOf(host) }
-        var portEditable by remember { mutableStateOf(port.toString()) }
+        var enabledEditable by remember { mutableStateOf(proxy.enabled) }
+        var hostEditable by remember { mutableStateOf(proxy.host) }
+        var portEditable by remember { mutableStateOf(proxy.port.toString()) }
+        var userEditable by remember { mutableStateOf(proxy.user) }
+        var passEditable by remember { mutableStateOf(proxy.pass) }
 
         var isValid by remember { mutableStateOf(true) }
 
-        fun validate(text: String, port: String): Boolean {
-            val portInt = port.toIntOrNull() ?: return false
+        fun validate(): Boolean {
+            val portInt = portEditable.toIntOrNull() ?: return false
 
             return try {
                 HttpUrl.Builder()
-                    .host(text)
+                    .host(hostEditable)
+                    .username(userEditable)
+                    .password(passEditable)
                     .port(portInt)
                     .scheme("http")
                     .build()
@@ -63,20 +70,24 @@ fun ProxyPrefDialog(modifier: Modifier = Modifier) {
             }
         }
 
-        LaunchedEffect(hostEditable, portEditable) {
+        LaunchedEffect(hostEditable, portEditable, userEditable, passEditable) {
             delay(500) // debounce
-            isValid = hostEditable.isNotBlank() && validate(hostEditable, portEditable)
+            isValid = hostEditable.isNotBlank() && validate()
         }
 
         DisposableEffect(Unit) {
             onDispose {
-                if (validate(hostEditable, portEditable)) {
+                if (validate()) {
                     Stuff.appScope.launch {
                         PlatformStuff.mainPrefs.updateData {
                             it.copy(
-                                customProxyEnabled = enabledEditable,
-                                proxyHost = hostEditable,
-                                proxyPort = portEditable.toInt()
+                                proxy = MainPrefs.ProxySettings(
+                                    enabled = enabledEditable,
+                                    host = hostEditable,
+                                    port = portEditable.toInt(),
+                                    user = userEditable,
+                                    pass = passEditable,
+                                )
                             )
                         }
                     }
@@ -102,7 +113,6 @@ fun ProxyPrefDialog(modifier: Modifier = Modifier) {
             verticalAlignment = Alignment.CenterVertically,
             modifier = Modifier
                 .fillMaxWidth()
-                .padding(horizontal = 8.dp, vertical = 8.dp)
         ) {
             PanoOutlinedTextField(
                 value = hostEditable,
@@ -132,6 +142,40 @@ fun ProxyPrefDialog(modifier: Modifier = Modifier) {
                 isError = !isValid,
                 modifier = Modifier
                     .width(120.dp)
+            )
+        }
+
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier
+                .fillMaxWidth()
+        ) {
+            PanoOutlinedTextField(
+                value = userEditable,
+                onValueChange = { userEditable = it },
+                label = {
+                    Text(stringResource(Res.string.username))
+                },
+                enabled = enabledEditable,
+                singleLine = true,
+                isError = !isValid,
+                modifier = Modifier
+                    .weight(1f)
+                    .padding(end = 8.dp)
+            )
+
+            PanoOutlinedTextField(
+                value = passEditable,
+                onValueChange = { passEditable = it },
+                label = {
+                    Text(stringResource(Res.string.password))
+                },
+                visualTransformation = PasswordVisualTransformation(),
+                enabled = enabledEditable,
+                singleLine = true,
+                isError = !isValid,
+                modifier = Modifier
+                    .weight(1f)
             )
         }
     }